requirements/prod.txt
fastapi==0.100.1
pydantic==1.10.11
uvicorn==0.23.2
boto3==1.28.17
PyYAML==6.0.1
langchain==0.0.255
openai
httpx==0.24.1
websockets==11.0.3